Working on do the train 1.0.0
Finally complete 'do the train'
Ready to main branch
