+++ lib/features/auth/auth_sign_in/presentation/controllers/sign_in_controller.dart
@@ -1,8 +1,5 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
-
 import 'package:fitflow/features/auth/auth_sign_in/domain/models/sign_in_state.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
-
 import 'package:fitflow/features/auth/auth_sign_in/domain/providers/auth_sign_in_domain_provider.dart';
 
 part 'sign_in_controller.g.dart';
