+++ lib/features/auth/auth_sign_out/data/repo/auth_sign_out_data_repo.dart
@@ -1,6 +1,4 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:supabase_flutter/supabase_flutter.dart';
-
 import 'package:fitflow/features/auth/auth_sign_out/data/repo/auth_sign_out_data_repo_impl.dart';
 
 class AuthSignOutDataRepo implements AuthSignOutDataRepoImpl {
