+++ lib/features/auth/auth_state_new/domain/models/app_user_state.dart
@@ -1,6 +1,4 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:equatable/equatable.dart';
-
 import 'package:fitflow/features/auth/auth_state_new/domain/models/app_user.dart';
 
 enum AppUserStateStatus { auth, unauth, unknown }
