+++ lib/features/auth/presentation/sign_in_page/sign_in_main_widget.dart
@@ -60,6 +60,5 @@ class SignInMainWidgetState extends ConsumerState<SignInMainWidget> {
         ),
       ],
     );
-    // );
   }
 }
