+++ lib/features/home/future_profile/domain/repo/home_update_indicators_repo.dart
@@ -1,4 +1,3 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:fitflow/features/home/future_profile/data/repo/home_update_indicators_repo.dart';
 import 'package:fitflow/features/home/future_profile/domain/repo/home_update_indicators_impl.dart';
