+++ lib/features/train/create_training_plan/presentation/view_done_plan/edit_day_in_plan/components/back_to_view_done_plan_button.dart
@@ -1,0 +1,53 @@
+import 'dart:math';
+
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+class BackToViewDonePlanButton extends StatelessWidget {
+  const BackToViewDonePlanButton({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Align(
+      alignment: Alignment.bottomCenter,
+      child: Padding(
+        padding: const EdgeInsets.only(
+          left: 39,
+          right: 39,
+          bottom: 35,
+        ),
+        child: Container(
+          decoration: BoxDecoration(
+              borderRadius: const BorderRadius.all(Radius.circular(99)),
+              gradient: LinearGradient(colors: [
+                Theme.of(context).colorScheme.secondary,
+                Theme.of(context).colorScheme.primary,
+              ], transform: const GradientRotation(pi / 4))),
+          child: ElevatedButton(
+              onPressed: () {
+                context.pop();
+              },
+              style: ButtonStyle(
+                  elevation: const WidgetStatePropertyAll(0),
+                  fixedSize: WidgetStatePropertyAll(
+                      Size(MediaQuery.of(context).size.width, 60)),
+                  backgroundColor:
+                      const WidgetStatePropertyAll(Colors.transparent)),
+              child: FittedBox(
+                child: Text(
+                  'Продолжить',
+                  textScaler: const TextScaler.linear(1),
+                  style: GoogleFonts.inter(
+                      color: Theme.of(context).colorScheme.onSecondary,
+                      fontSize: 16,
+                      fontWeight: FontWeight.w700),
+                ),
+              )),
+        ),
+      ),
+    );
+  }
+}
